requirements.txt
# Dependencies for CAN 2025 Chatbot Pipeline

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data processing
python-dateutil==2.8.2

# RAG and Vector Database
chromadb==0.4.22
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
groq==0.4.1

# Embeddings (Open Source - Gratuit)
sentence-transformers==2.2.2
torch==2.1.0

# Web Interface
streamlit>=1.52.0

# Logging
colorlog==6.8.2

# Environment variables
python-dotenv==1.0.0
